Generate Contacts + getContacts API
