+++ src/routes/api.routes.js
@@ -5,7 +5,6 @@ const router = express.Router();
 
 const userRoutes = require('../users/user.controller');
 const profilesRoutes = require('../profiles/profiles.controller');
-const contactsRoutes = require('../contacts/contact.controller');
 
 router.use('/secured', security.checkAuthenticated);
 
@@ -25,6 +24,6 @@ router.post('/public/forgot-password', userRoutes.forgottenPassord);
 router.get('/secured/users', userRoutes.getUsers);
 router.put('/secured/users', userRoutes.updateUser);
 
-router.get('/secured/users/contacts', contactsRoutes.getContacts);
+router.get('/secured/users/contacts', userRoutes.getContacts);
 
 module.exports = router;
