+++ src/users/user.controller.js
@@ -83,10 +83,23 @@ function getUsers(req, res) {
     });
 }
 
+function getContacts(req, res) {
+  userServices.getContacts(req.user.phone)
+    .then((contacts) => {
+      res.status(200).json(contacts);
+    })
+    .catch((e) => {
+      const message = e.message || 'Internal server error';
+      const status = e.message ? 400 : 500;
+      res.status(status).json({ message });
+    });
+}
+
 module.exports = {
   signIn,
   logIn,
   forgottenPassord,
   getUsers,
   updateUser,
+  getContacts,
 };
